requirements.txt
# Core LLM
groq==0.11.0

# LangChain (modern tool calling)
langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.32

# Prompt + orchestration
pydantic==1.10.13

# Utilities
python-dotenv==1.0.1
jinja2==3.1.3

# JSON safety
jsonschema==4.22.0

# HTTP (Groq dependency)
httpx==0.27.0